overhaul front end to keep all the content in App.tsx
