--- autonomousprice/front/.prettierignore
@@ -1,5 +1,0 @@
-reportWebVitals.js
-setupTests.js
-
-dist
-*.min.js
