--- autonomousprice/front/.prettierrc
@@ -1,6 +1,0 @@
-{
-    "trailingComma": "es5",
-    "tabWidth": 4,
-    "semi": true,
-    "singleQuote": false
-}
