+++ autonomousprice/front/package.json
@@ -1,52 +1,86 @@
 {
-    "name": "autonomous-price-front",
-    "version": "0.1.0",
-    "private": true,
-    "dependencies": {
-        "@massalabs/massa-web3": "^1.22.0",
-        "@massalabs/react-ui-kit": "^0.0.4-dev",
-        "@massalabs/wallet-provider": "^1.2.0",
-        "@types/node": "^16.18.37",
-        "@types/react": "^18.2.14",
-        "@types/react-dom": "^18.2.6",
-        "react": "^18.2.0",
-        "react-dom": "^18.2.0",
-        "react-router-dom": "^6.14.2",
-        "react-scripts": "5.0.1",
-        "typescript": "^4.9.5",
-        "web-vitals": "^2.1.4"
-    },
-    "scripts": {
-        "start": "react-scripts start",
-        "build": "react-scripts build",
-        "test": "react-scripts test",
-        "eject": "react-scripts eject",
-        "prettier:fix": "prettier --write \"**/*.js\""
-    },
-    "eslintConfig": {
-        "extends": [
-            "react-app",
-            "react-app/jest"
-        ]
-    },
-    "browserslist": {
-        "production": [
-            ">0.2%",
-            "not dead",
-            "not op_mini all"
-        ],
-        "development": [
-            "last 1 chrome version",
-            "last 1 firefox version",
-            "last 1 safari version"
-        ]
-    },
-    "devDependencies": {
-        "@testing-library/jest-dom": "^5.16.5",
-        "@testing-library/react": "^13.4.0",
-        "@testing-library/user-event": "^13.5.0",
-        "@types/jest": "^27.5.2",
-        "prettier": "^2.8.8",
-        "tailwindcss": "^3.3.2"
-    }
+  "name": "autonomous-price-front",
+  "version": "1.0.0",
+  "target": "esnext",
+  "description": "massa's create-react-app template",
+  "author": "Massa Labs <info@massa.net>",
+  "contributors": [],
+  "license": "MIT",
+  "private": false,
+  "keywords": [
+    "web3",
+    "js",
+    "ts",
+    "react",
+    "massa",
+    "web"
+  ],
+  "dependencies": {
+    "@massalabs/massa-web3": "^1.22.0",
+    "@massalabs/react-ui-kit": "^0.0.3",
+    "@massalabs/wallet-provider": "^1.2.0",
+    "@testing-library/jest-dom": "^5.16.5",
+    "@testing-library/react": "^13.4.0",
+    "@testing-library/user-event": "^14.4.3",
+    "@types/jest": "^29.2.3",
+    "@types/node": "^18.11.9",
+    "@types/react": "^18.0.25",
+    "@types/react-dom": "^18.0.9",
+    "assert": "^2.0.0",
+    "browserify-fs": "^1.0.0",
+    "browserify-zlib": "^0.2.0",
+    "buffer": "^6.0.3",
+    "crypto-browserify": "^3.12.0",
+    "events": "^3.3.0",
+    "https-browserify": "^1.0.0",
+    "madge": "^5.0.1",
+    "os-browserify": "^0.3.0",
+    "path": "^0.12.7",
+    "process": "^0.11.10",
+    "querystring-es3": "^0.2.1",
+    "react": "^18.2.0",
+    "react-app-rewired": "^2.2.1",
+    "react-dom": "^18.2.0",
+    "react-scripts": "5.0.1",
+    "readable-stream": "^4.2.0",
+    "stream": "^0.0.2",
+    "stream-http": "^3.2.0",
+    "timers-browserify": "^2.0.12",
+    "typescript": "^4.9.3",
+    "url": "^0.11.0",
+    "util": "^0.12.5",
+    "util-browser": "^0.0.2",
+    "web-vitals": "^3.1.0",
+    "webpack-cli": "^5.0.0"
+  },
+  "devDependencies": {
+    "process": "^0.11.10"
+  },
+  "scripts": {
+    "start": "react-app-rewired start",
+    "build": "react-app-rewired build",
+    "test": "react-app-rewired test",
+    "eject": "react-scripts eject",
+    "fix-lint": "tslint -c tslint.json '**/*.ts' --fix"
+  },
+  "eslintConfig": {
+    "extends": [
+      "react-app",
+      "react-app/jest"
+    ]
+  },
+  "browserslist": {
+    "production": [
+      "chrome >= 67",
+      "edge >= 79",
+      "firefox >= 68",
+      "opera >= 54",
+      "safari >= 14"
+    ],
+    "development": [
+      "last 1 chrome version",
+      "last 1 firefox version",
+      "last 1 safari version"
+    ]
+  }
 }
