+++ autonomousprice/front/src/App.css
@@ -1,7 +1,92 @@
-@tailwind base;
-@tailwind components;
-@tailwind utilities;
+/* Setup the background color */
+body {
+  background-image: url("../public/background.png");
+  background-repeat: no-repeat;
+  background-size: cover;
+  padding-bottom: 4rem;
+  overflow-y: auto;
+  height: 100vh;
+  overflow: hidden;
+}
 
-.App {
-    text-align: center;
+.bodyContent {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  box-sizing: border-box;
+  height: 90vh;
+  align-self: center;
+}
+
+.wrapper {
+  width: 100%;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  box-sizing: border-box;
+  justify-content: center;
+}
+
+
+.innerWrapper {
+  width: 100%;
+  display: flex;
+  align-items: center;
+  box-sizing: border-box;
+  justify-content: flex-start;
+  padding-top: 1.5rem;
+}
+
+.invisibleDiv {
+  visibility: hidden;
+}
+
+input {
+  display: block;
+  padding: 10px;
+  box-sizing: border-box;
+  border: 1px solid #ddd;
+  border-radius: 4px;
+  font-weight: 500;
+  font-size: 1.1em;
+  margin-right: 2rem;
+}
+
+/* Setup the bodyText text */
+
+.messageToDisplay {
+  color: rgb(0, 0, 0);
+  font-size: 2em;
+  font-weight: bold;
+  text-align: center;
+  padding-bottom: 1rem;
+}
+
+h3 {
+  font-size: 1.2em;
+  font-weight: 500;
+  text-align: center;
+}
+
+h4 {
+  font-size: 1.1em;
+  font-weight: 500;
+  text-align: center;
+}
+
+button {
+  width: 35%;
+  padding: 12px 16px;
+  border-radius: 8px;
+  border: none;
+  cursor: pointer;
+  text-align: center;
+  font-size: 1.3em;
+  color: rgb(255, 255, 255);
+  background-color: rgb(159, 26, 26);
+}
+
+button:hover {
+  background-color: rgb(73, 47, 47);
 }
