--- autonomousprice/front/src/components/body.tsx
@@ -1,11 +1,0 @@
-import ContractInteraction from "./contractInteraction";
-
-export default function Body() {
-    return (
-        <div className="body flex-col justify-center align-middle">
-            <div className="flex-col">
-                <ContractInteraction />
-            </div>
-        </div>
-    );
-}
