--- autonomousprice/front/src/const/index.ts
@@ -1,8 +1,0 @@
-export const MASSA_EXEMPLE = {
-    TITLE: "Massa Autonomous Price",
-    DESCRIPTION: "This is a simple dApp showcasing Massa's Autonomous Smart Contract feature.",
-    TAB: "Massa - Autonomous Price example",
-};
-
-export const CONTRACT_ADDRESS =
-    "AS12tdw4onaZqFkSTHqSFWaJZxSQST2EWU7xQoXmFETDkSEiZPLjs";
