+++ autonomousprice/front/src/setupTests.ts
